requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
distlib==0.3.1
Django==3.1.7
django-braces==1.14.0
django-model-utils==4.1.1
django-star-ratings==0.9.1
djangorestframework==3.12.4
drf-writable-nested==0.6.2
filelock==3.0.12
gunicorn==20.1.0
httpie==2.4.0
idna==2.10
pbr==5.5.1
Pillow==8.2.0
psycopg2==2.8.6
Pygments==2.8.1
PySocks==1.7.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
stevedore==3.3.0
swapper==1.1.2.post1
urllib3==1.26.4
virtualenv==20.4.3
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
